[Feature] Core features implemented
User now can scan barcode using CameraActivity.kt and pass the scanned value to our repository network part to check if there is any information on the REST API with that number.

The server using ISBN-13 first then trying with ISBN-10 code. If there is, it checks the different book editions issuing year and pass that back to fill AutoCompleteTextView options and also grabs a cover photo of the scanned book. All information passed to the EditText views too.

When user clicks on the save FAB, we check if all requested fields a good and save all information to our database.